making client for display all clients(loan details) of service provider and change enum for all loans

+++ service-loans/src/main/java/com/capitaworld/service/loans/domain/fundseeker/LoanApplicationMaster.java
@@ -58,7 +58,7 @@ public class LoanApplicationMaster implements Serializable {
 	private String name;
 
 	@Column(name = "product_id")
-	private Integer productId;
+	private Long productId;
 
 	private Integer tenure;
 
@@ -155,11 +155,12 @@ public class LoanApplicationMaster implements Serializable {
 		this.name = name;
 	}
 
-	public Integer getProductId() {
+	
+	public Long getProductId() {
 		return productId;
 	}
 
-	public void setProductId(Integer productId) {
+	public void setProductId(Long productId) {
 		this.productId = productId;
 	}
